refactor: remove 'useXcmTokenDetails' hook
