+++ src/hooks/index.ts
@@ -32,6 +32,5 @@ export * from './wallet/useEvmWallet';
 export * from './xcm/useXcmAssets';
 export * from './xcm/useXcmBridge';
 export * from './xcm/useXcmTokenTransfer';
-export * from './xcm/useXcmTokenDetails';
 export * from './xcm/useXcmEvm';
 export * from './ui/useTooltip';
